BruteForce_ai_part/requirements.txt
# YouTube transcript extraction
youtube-transcript-api==1.5.0

# Gemini API client (Google Generative AI SDK)
google-generativeai==0.4.1

# Text-to-Speech
gTTS==2.3.2

# Optional: text processing and utilities
numpy==1.26.0
pandas==2.1.0

# Optional: for handling audio files
pydub==0.25.1
